+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to the USANA stock manager will be documented in this file.
 ## 0.1.1
 
 - Viewing and editing loan records
-- Viewing loan summary
+- Viewing loan summary which can redirect to editing loan records
 - Fuzzy search for stock view
 
 ## 0.1.0